Execute analyses only when ready
